stats: move statistics utilities to records
